+++ KMaP/src/commonMain/kotlin/com/rafambn/kmap/core/MapState.kt
@@ -1,7 +1,6 @@
 package com.rafambn.kmap.core
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -10,9 +9,7 @@ import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.unit.Density
 import com.rafambn.kmap.mapProperties.MapProperties
 import com.rafambn.kmap.mapProperties.border.MapBorderType
-import com.rafambn.kmap.mapProperties.border.OutsideTilesType
 import com.rafambn.kmap.mapProperties.coordinates.MapCoordinatesRange
-import com.rafambn.kmap.tiles.TileSpecs
 import com.rafambn.kmap.utils.CanvasDrawReference
 import com.rafambn.kmap.utils.CanvasPosition
 import com.rafambn.kmap.utils.DifferentialScreenOffset
@@ -25,7 +22,6 @@ import com.rafambn.kmap.utils.loopInRange
 import com.rafambn.kmap.utils.rotate
 import com.rafambn.kmap.utils.toIntFloor
 import com.rafambn.kmap.utils.toRadians
-import kotlin.math.pow
 
 @Composable
 fun rememberMapState(
@@ -63,22 +59,13 @@ class MapState(
         }
     val drawReference
         get() = cameraState.rawPosition.toCanvasDrawReference()
-    val zoomLevel
+    private val zoomLevel
         get() = cameraState.zoom.toIntFloor()
-    val magnifierScale
+    private val magnifierScale
         get() = cameraState.zoom - zoomLevel + 1F
 
-    val visibleTiles by derivedStateOf {
-        getVisibleTilesForLevel(
-            getBoundingBox(),
-            zoomLevel,
-            mapProperties.outsideTiles,
-            mapProperties.mapCoordinatesRange
-        )
-    }
-
     //Utility functions
-    private fun getBoundingBox(): BoundingBox {
+    fun getBoundingBox(): BoundingBox {
         return BoundingBox(
             ScreenOffset.Zero.toCanvasPosition(),
             ScreenOffset(cameraState.canvasSize.x, 0F).toCanvasPosition(),
@@ -169,81 +156,6 @@ class MapState(
         return CanvasPosition(
             horizontal * mapCoordinatesRange.longitude.getOrientation(),
             vertical * mapCoordinatesRange.latitude.getOrientation()
-        )
-    }
-
-    private fun CanvasPosition.applyInverseOrientation(mapCoordinatesRange: MapCoordinatesRange): CanvasPosition {
-        return CanvasPosition(
-            horizontal * mapCoordinatesRange.longitude.getOrientation() * -1,
-            vertical * mapCoordinatesRange.latitude.getOrientation() * -1
-        )
-    }
-
-    private fun getVisibleTilesForLevel(
-        viewPort: BoundingBox,
-        zoomLevel: Int,
-        outsideTilesType: OutsideTilesType,
-        coordinatesRange: MapCoordinatesRange
-    ): List<TileSpecs> {
-        val topLeftTile = getXYTile(
-            viewPort.topLeft.applyInverseOrientation(coordinatesRange),
-            zoomLevel,
-            coordinatesRange
-        )
-        val topRightTile = getXYTile(
-            viewPort.topRight.applyInverseOrientation(coordinatesRange),
-            zoomLevel,
-            coordinatesRange
-        )
-        val bottomRightTile = getXYTile(
-            viewPort.bottomRight.applyInverseOrientation(coordinatesRange),
-            zoomLevel,
-            coordinatesRange
-        )
-        val bottomLeftTile = getXYTile(
-            viewPort.bottomLeft.applyInverseOrientation(coordinatesRange),
-            zoomLevel,
-            coordinatesRange
-        )
-        val horizontalTileIntRange =
-            IntRange(
-                minOf(topLeftTile.first, bottomRightTile.first, topRightTile.first, bottomLeftTile.first),
-                maxOf(topLeftTile.first, bottomRightTile.first, topRightTile.first, bottomLeftTile.first)
-            )
-        val verticalTileIntRange =
-            IntRange(
-                minOf(topLeftTile.second, bottomRightTile.second, topRightTile.second, bottomLeftTile.second),
-                maxOf(topLeftTile.second, bottomRightTile.second, topRightTile.second, bottomLeftTile.second)
-            )
-
-        val visibleTileSpecs = mutableListOf<TileSpecs>()
-        if (outsideTilesType == OutsideTilesType.NONE) {
-            for (x in horizontalTileIntRange)
-                for (y in verticalTileIntRange) {
-                    var xTile: Int
-                    if (x < 0 || x > 2F.pow(zoomLevel) - 1)
-                        continue
-                    else
-                        xTile = x
-                    var yTile: Int
-                    if (y < 0 || y > 2F.pow(zoomLevel) - 1)
-                        continue
-                    else
-                        yTile = y
-                    visibleTileSpecs.add(TileSpecs(zoomLevel, xTile, yTile))
-                }
-        } else {
-            for (x in horizontalTileIntRange)
-                for (y in verticalTileIntRange)
-                    visibleTileSpecs.add(TileSpecs(zoomLevel, x, y))
-        }
-        return visibleTileSpecs
-    }
-
-    private fun getXYTile(position: CanvasPosition, zoomLevel: Int, mapSize: MapCoordinatesRange): Pair<Int, Int> {
-        return Pair(
-            ((position.horizontal - mapSize.longitude.getMin()) / mapSize.longitude.span * (1 shl zoomLevel)).toIntFloor(),
-            ((position.vertical - mapSize.latitude.getMin()) / mapSize.latitude.span * (1 shl zoomLevel)).toIntFloor()
         )
     }
