+++ KMaP/src/commonMain/kotlin/com/rafambn/kmap/tiles/TileRenderResult.kt
@@ -1,6 +1,6 @@
 package com.rafambn.kmap.tiles
 
-interface TileRenderResult {
+sealed interface TileRenderResult {
     data class Success(val tile: Tile) : TileRenderResult
     data class Failure(val specs: TileSpecs) : TileRenderResult
 }
